[5332] Create RecommendationsBaseController
- Create RecommendationsBaseController which all othe recommendations
  controllers can inherit from
- Move all shared helper methods and before_action into base controller
