+++ app/views/bulk_update/recommendations_checks/show.html.erb
@@ -1,4 +1,4 @@
-<%= render PageTitle::View.new(text: "Check who you’ll recommend for #{qts_or_eyts(@recommendations_upload.awardable_rows)}") %>
+<%= render PageTitle::View.new(text: "Check who you’ll recommend for #{qts_or_eyts(recommendations_upload.awardable_rows)}") %>
 
 <%= content_for(:breadcrumbs) do %>
   <%= render GovukComponent::BackLinkComponent.new(text: "Home", href: root_path) %>
@@ -9,23 +9,23 @@
     <h1 class="govuk-heading-l">
       <span class="govuk-caption-xl"><%= current_user.organisation.name %></span>
       Check who you’ll recommend for
-      <%= qts_or_eyts(@recommendations_upload.awardable_rows) %>
+      <%= qts_or_eyts(recommendations_upload.awardable_rows) %>
     </h1>
 
-    <% @recommendations_upload.awardable_rows.includes(:trainee).tap do |awardable_rows| %>
+    <% recommendations_upload.awardable_rows.includes(:trainee).tap do |awardable_rows| %>
       <% if awardable_rows.count <= @table_row_limit %>
         <%= render partial: "table", locals: {
           rows: awardable_rows,
-          caption: "Trainees you’ll recommend for #{qts_or_eyts(@recommendations_upload.awardable_rows)}"
+          caption: "Trainees you’ll recommend for #{qts_or_eyts(recommendations_upload.awardable_rows)}"
         } %>
       <% else %>
-        <%= govuk_accordion(id: "recommendations-upload-#{@recommendations_upload.id}",
+        <%= govuk_accordion(id: "recommendations-upload-#{recommendations_upload.id}",
                             html_attributes: {
                               "data-i18n.show-all-sections" => "Show all trainees you’ll recommend",
                               "data-i18n.hide-all-sections" => "Hide all trainees you’ll recommend",
                             }) do |accordion|
           awardable_rows.each_slice(@table_row_limit).with_index do |(*rows), index|
-            heading_text = "Trainees #{@table_row_limit*index + 1} to #{@table_row_limit*index + rows.count} you’ll recommend for #{qts_or_eyts(@recommendations_upload.awardable_rows)}"
+            heading_text = "Trainees #{@table_row_limit*index + 1} to #{@table_row_limit*index + rows.count} you’ll recommend for #{qts_or_eyts(recommendations_upload.awardable_rows)}"
             accordion.section(heading_text: heading_text) do
               render partial: "table", locals: { rows: rows, caption: heading_text }
             end
@@ -41,13 +41,13 @@
   <div class="govuk-grid-column-full">
     <div class="js-stick-at-bottom-when-scrolling">
       <%= register_form_with url: bulk_update_recommendations_upload_recommendations_path, method: :post, local: true do |f| %>
-        <%= f.govuk_submit("Recommend #{pluralize(@awardable_rows_count, "trainee")} for #{qts_or_eyts(@recommendations_upload.awardable_rows)}", id: "recommend") %>
+        <%= f.govuk_submit("Recommend #{pluralize(awardable_rows_count, "trainee")} for #{qts_or_eyts(recommendations_upload.awardable_rows)}", id: "recommend") %>
       <% end %>
       <div class="govuk-body">
-        <%= govuk_link_to("Change who you’ll recommend", edit_bulk_update_recommendations_upload_path(@recommendations_upload)) %>
+        <%= govuk_link_to("Change who you’ll recommend", edit_bulk_update_recommendations_upload_path(recommendations_upload)) %>
       </div>
       <div class="govuk-body">
-        <%= govuk_link_to("Cancel bulk recommending trainees", cancel_bulk_update_recommendations_upload_path(@recommendations_upload)) %>
+        <%= govuk_link_to("Cancel bulk recommending trainees", cancel_bulk_update_recommendations_upload_path(recommendations_upload)) %>
       </div>
     </div>
   </div>
