+++ app/views/bulk_update/recommendations_uploads/confirmation.html.erb
@@ -5,13 +5,13 @@
     <div class="govuk-panel govuk-panel--confirmation qts-submission-success-panel govuk-!-margin-bottom-6">
       <h1 class="govuk-panel__title">
         <%= pluralize(awardable_rows_count, "trainee") %> recommended for
-        <%= qts_or_eyts(@recommendations_upload.awardable_rows) %>
+        <%= qts_or_eyts(recommendations_upload.awardable_rows) %>
       </h1>
     </div>
 
     <p class="govuk-body">
       The Department for Education will award
-      <%= qts_or_eyts(@recommendations_upload.awardable_rows) %> within 3
+      <%= qts_or_eyts(recommendations_upload.awardable_rows) %> within 3
       working days.
     </p>
